// src/treemap/src/treemap/model/base-model.d.ts
export interface BorderModel {

    /**
     * The color of the border that accepts value in hex and rgba as a valid CSS color string.
     * @default '#808080'
     */
    color?: string;

    /**
     * The width of the border in pixels.
     * @default 0
     */
    width?: number;

}
// ...
export interface SelectionSettingsModel {

    /**
     * To enable or disable the selection
     * @default false
     */
    enable?: boolean;

    /**
     * To specifies the selection color
     * @default '#808080'
     */
    fill?: string;

    /**
     * To specified the opacity of color.
     * @default '0.5'
     */
    opacity?: string;

    /**
     * To specifies the border
     */
    border?: BorderModel;

    /**
     * To specifies the selection mode.
     * @default 'Item'
     */
    mode?: SelectionMode;

}

/**
 * Interface for a class HighlightSettings
 */
export interface HighlightSettingsModel {

    /**
     * To enable or disable the highlight.
     * @default false
     */
    enable?: boolean;

    /**
     * To specifies the highlight color.
     * @default '#808080'
     */
    fill?: string;

    /**
     * To specified the opacity of color.
     * @default '0.5'
     */
    opacity?: string;

    /**
     * To specifies the border
     */
    border?: BorderModel;

    /**
     * To specifies the highlight mode.
     * @default 'Item'
     */
    mode?: HighLightMode;

}